fix(organization): decouple client and server permission checks (#4707)
