+++ packages/better-auth/src/plugins/organization/client.ts
@@ -12,11 +12,29 @@ import { type AccessControl, type Role } from "../access";
 import type { BetterAuthClientPlugin } from "../../client/types";
 import { organization } from "./organization";
 import { useAuthQuery } from "../../client";
-import { defaultStatements, adminAc, memberAc, ownerAc } from "./access";
-import { clientSideHasPermission } from "./has-permission";
+import {
+	defaultStatements,
+	adminAc,
+	memberAc,
+	ownerAc,
+	defaultRoles,
+} from "./access";
 import type { FieldAttribute } from "../../db";
 import type { BetterAuthOptions, BetterAuthPlugin } from "../../types";
 import type { OrganizationOptions } from "./types";
+import type { HasPermissionBaseInput } from "./permission";
+import { hasPermissionFn } from "./permission";
+
+/**
+ * Using the same `hasPermissionFn` function, but without the need for a `ctx` parameter or the `organizationId` parameter.
+ */
+export const clientSideHasPermission = (input: HasPermissionBaseInput) => {
+	const acRoles: {
+		[x: string]: Role<any> | undefined;
+	} = input.options.roles || defaultRoles;
+
+	return hasPermissionFn(input, acRoles);
+};
 
 interface OrganizationClientOptions {
 	ac?: AccessControl;
