+++ packages/better-auth/src/plugins/organization/permission.ts
@@ -1,0 +1,53 @@
+import type { Role } from "../access";
+import type { OrganizationOptions } from "./types";
+
+export const hasPermissionFn = (
+	input: HasPermissionBaseInput,
+	acRoles: {
+		[x: string]: Role<any> | undefined;
+	},
+) => {
+	if (!input.permissions && !input.permission) return false;
+
+	const roles = input.role.split(",");
+	const creatorRole = input.options.creatorRole || "owner";
+	const isCreator = roles.includes(creatorRole);
+
+	const allowCreatorsAllPermissions = input.allowCreatorAllPermissions || false;
+	if (isCreator && allowCreatorsAllPermissions) return true;
+
+	for (const role of roles) {
+		const _role = acRoles[role as keyof typeof acRoles];
+		const result = _role?.authorize(input.permissions ?? input.permission);
+		if (result?.success) {
+			return true;
+		}
+	}
+	return false;
+};
+
+export type PermissionExclusive =
+	| {
+			/**
+			 * @deprecated Use `permissions` instead
+			 */
+			permission: { [key: string]: string[] };
+			permissions?: never;
+	  }
+	| {
+			permissions: { [key: string]: string[] };
+			permission?: never;
+	  };
+
+export let cacheAllRoles = new Map<
+	string,
+	{
+		[x: string]: Role<any> | undefined;
+	}
+>();
+
+export type HasPermissionBaseInput = {
+	role: string;
+	options: OrganizationOptions;
+	allowCreatorAllPermissions?: boolean;
+} & PermissionExclusive;
